pyproject.toml
[project]
name = "xkcd-bot"
version = "0.1.0"
description = "reddit bot for xkcd comic"
authors = [
    {name = "Samarth Rawat",email = "samarthrawat1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "praw (>=7.8.1,<8.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "flask (>=3.1.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
